+++ example/src/App.tsx
@@ -43,8 +43,11 @@ import CashAppScreen from './screens/CashAppScreen';
 import PaymentSheetDeferredIntentScreen from './screens/PaymentSheetDeferredIntentScreen';
 import PaymentSheetDeferredIntentMultiStepScreen from './screens/PaymentSheetDeferredIntentMultiStepScreen';
 import EmbeddedPaymentElementScreen from './screens/EmbeddedPaymentElementScreen';
+import EmbeddedPaymentElementImmediateActionScreen from './screens/EmbeddedPaymentElementImmediateActionScreen';
+import EmbeddedPaymentElementConfirmScreen from './screens/EmbeddedPaymentElementConfirmScreen';
 import CustomerSheetScreen from './screens/CustomerSheetScreen';
 import RevolutPayScreen from './screens/RevolutPayScreen';
+import type { EmbeddedPaymentElementResult } from '@stripe/stripe-react-native';
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
 
@@ -89,6 +92,10 @@ export type RootStackParamList = {
   PaymentSheetDeferredIntentScreen: undefined;
   PaymentSheetDeferredIntentMultiStepScreen: undefined;
   EmbeddedPaymentElementScreen: undefined;
+  EmbeddedPaymentElementImmediateActionScreen: undefined;
+  EmbeddedPaymentElementConfirmScreen: {
+    confirm: () => Promise<EmbeddedPaymentElementResult>;
+  };
   CustomerSheetScreen: undefined;
   RevolutPayScreen: undefined;
 };
@@ -168,6 +175,14 @@ export default function App() {
           <Stack.Screen
             name="EmbeddedPaymentElementScreen"
             component={EmbeddedPaymentElementScreen}
+          />
+          <Stack.Screen
+            name="EmbeddedPaymentElementImmediateActionScreen"
+            component={EmbeddedPaymentElementImmediateActionScreen}
+          />
+          <Stack.Screen
+            name="EmbeddedPaymentElementConfirmScreen"
+            component={EmbeddedPaymentElementConfirmScreen}
           />
           <Stack.Screen
             name="PaymentsUICustomScreen"