+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **Feature**
 - Added rowSelectionBehavior to `EmbeddedPaymentElementConfiguration` with `immediateAction` option
+- Added `flatWithChevron` to `AppearanceParams.embeddedPaymentElement.rowConfig.style`
 
 ## 0.48.0 - 2025-06-11